Add backend support in TripDashboard
