+++ src/screens/TripsDashboard/types/trip-dashboard.query.ts
